requirements.txt
langchain
streamlit
python-dotenv
datetime
pathlib
agentql
#-e.